Rethinking of ImageIOHandlerPlugin API.
As we use mimeTypes for plugin detection, there is no need to do magic
checks manually as it is already done by mime magic rules.

So, we do not pass QIODevice pointer to the plugin and do not peek magic
number in factory.

In case if user forces to use specific mimetype, ok we fail on header
validation.