Move to tl::expected.
Just a proof of concept (working), however it shows some flaws:
First, poor implementation of the given tl::expected -
tl::expected<void, Error> is not working as expected (i.e. is not
assignable)

Second, TextureIOResult should be rewritten, it should emulate
expected<void, Error> where Error is a simple enum. String conversion
should be moved out of the TextureIOError class because this is the only
reason for the whole class, toString(Error) (where Error is a enum)
should be sufficient)